Add support for loading PSX BINs
This works weirdly with the way the program currently acts. Since it extracts files to the temp folder when the user starts the modding process, we must keep the converted copy saved somewhere so we don't repeatedly convert and waste a billion years on that. Maybe some PS1- (and maybe GC?) specific changes could be made at some point.