Rewrite how modders generally work, using abstract classes
Using an abstract class instead of reflection makes the program much easier to debug and code for, since it forces certain functions to be present. It also makes it so you have way less copy-pasted variables and enums, less casting and less copy-pasted functions since it's all done more automagically.

Also, classes are reference types (unlike structs), so using the ref keyword is redundant. It should only be used on value types that you want to change inside a function but keep using outside of it.